\name{operatorCalc-package}
\alias{operatorCalc-package}
\alias{operatorCalc}
\docType{package}
\title{
Calculus for some operators on multvariate function space
}
\description{
Aritmetic methods for a class of piecewise polynomial matrix functions. Aritmetic methods for the class of so-called lattice operators on function space.
}
\details{
\tabular{ll}{
Package: \tab operatorCalc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-05\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Bo Markussen <bomar@math.ku.dk>
}
%% \references{
%% ~~ Literature or other references for background information ~~
%% }
\keyword{
  package
  math
}
\seealso{
Package for spline functions \code{\link{splines}}.
}
\examples{
mesh(seq(0,1,length.out=11))
}
